move sql models to designated folder
